Delete unused file (index.js), change name of folder api to auth (more suitable), simplifying the user auth function, routes optimization in routers (userRouter and contactRouter) , correcting the message when choosing non-existent route
