+++ auth/user.auth.js
@@ -1,0 +1,21 @@
+import passport from 'passport';
+
+const auth = async (req, res, next) => {
+  passport.authenticate('jwt', { session: false }, (err, user) => {
+    const token = req.get('Authorization').replace('Bearer ', '');
+
+    if (!user || err || !token || token !== user.token) {
+      return res.status(401).json({
+        status: 'error',
+        code: 401,
+        message: 'Not authorized',
+        data: 'Unauthorized',
+      });
+    }
+
+    req.user = user;
+    next();
+  })(req, res, next);
+};
+
+export default auth;
